+++ include/toltrig/constants.hpp
@@ -4,6 +4,7 @@ namespace toltrig {
 
 constexpr double pi = 3.141592653589793238462643383279502884;
 constexpr double half_pi = pi / 2.0;
+constexpr double quarter_pi = pi / 4.0;
 constexpr double two_pi = 2.0 * pi;
 
 }  // namespace toltrig